+++ FitCook/Modules/Component/FCDosageView/FCFractionalView.m
@@ -63,7 +63,7 @@
         _labSeparator.text = nil;
     } else {
         _labNumerator.text = [NSString stringWithFormat:@"%ld",numerator];
-        _labDenominator.text = [NSString stringWithFormat:@"%ld",denominator];
+        _labDenominator.text = [NSString stringWithFormat:@"%ld",(long)denominator];
         _labSeparator.text = @"/";
     }
 }
